requirements.txt
asgiref==3.7.2
Brotli==1.0.9
cffi==1.15.1
cryptography==41.0.1
dj-database-url==2.0.0
Django==3.2.20
django-rest-knox==4.2.0
djangorestframework==3.14.0
gunicorn==20.1.0
Pillow==9.5.0
psycopg2-binary==2.9.6
pycparser==2.21
python-dotenv==0.21.1
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.6.3
tzdata==2023.3
whitenoise==6.4.0
djangorestframework-simplejwt==5.2.0
requests==2.28.1